requirements.txt
facenet_pytorch==0.1.0
torch==1.2.0
torchvision==0.4.0
numpy==1.16.5
Pillow==6.1.0
joblib==0.13.2
scikit_learn==0.21.3
scikit-image
Werkzeug==0.15.2
